Refactor Document Management System
- Removed DocumentService.Commands.cs and DocumentService.Queries.cs files to streamline document handling.
- Deleted CloudinaryStorageService.cs as cloud storage functionality is no longer required.
- Updated project structure by removing the old Services project and introducing a new Application, Domain, and Infrastructure layers.
- Added new entities: AppUser, Document, and Subject to the Domain layer.
- Implemented AistudyHubDbContext for database interactions in the Infrastructure layer.
- Created new WebAPI project for handling HTTP requests and responses.
- Established JWT authentication with RefreshToken management.
- Updated database schema to include new tables and relationships for documents, users, and chat sessions.
- Added Swagger for API documentation and testing.
- Configured appsettings for development environment including database connection and Cloudinary settings.
