--- Prn212.AIStudyHub.Services/Prn212.AIStudyHub.Services.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="CloudinaryDotNet" Version="1.27.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.*" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Prn212.AIStudyHub.DataAccess\Prn212.AIStudyHub.DataAccess.csproj" />
-  </ItemGroup>
-
-</Project>
