+++ Prn212.AIStudyHub.WPF/Prn212.AIStudyHub.WPF.csproj
@@ -2,16 +2,11 @@
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net8.0-windows</TargetFramework>
+    <TargetFramework>net10.0-windows</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <UseWPF>true</UseWPF>
   </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Prn212.AIStudyHub.Services\Prn212.AIStudyHub.Services.csproj" />
-    <ProjectReference Include="..\Prn212.AIStudyHub.DataAccess\Prn212.AIStudyHub.DataAccess.csproj" />
-  </ItemGroup>
 
   <ItemGroup>
     <None Update="appsettings.json">
