+++ src/PRN212.AIStudyHub.Domain/Entities/AppUser.cs
@@ -1,0 +1,15 @@
+namespace PRN212.AIStudyHub.Domain.Entities;
+
+public partial class AppUser
+{
+    public int Id { get; set; }
+    public string Email { get; set; } = null!;
+    public string PasswordHash { get; set; } = null!;
+    public string FirstName { get; set; } = null!;
+    public string LastName { get; set; } = null!;
+    public string Role { get; set; } = null!;
+    public bool IsActive { get; set; }
+    public DateTime CreatedAt { get; set; }
+    public DateTime? UpdatedAt { get; set; }
+    public virtual ICollection<Document> Documents { get; set; } = new List<Document>();
+}
