+++ src/PRN212.AIStudyHub.Domain/Entities/Subject.cs
@@ -1,0 +1,10 @@
+namespace PRN212.AIStudyHub.Domain.Entities;
+
+public partial class Subject
+{
+    public int Id { get; set; }
+    public string Name { get; set; } = null!;
+    public string? Description { get; set; }
+    public DateTime CreatedAt { get; set; }
+    public virtual ICollection<Document> Documents { get; set; } = new List<Document>();
+}
