+++ src/PRN212.AIStudyHub.WebAPI/Properties/serviceDependencies.json
@@ -1,0 +1,8 @@
+{
+  "dependencies": {
+    "secrets1": {
+      "type": "secrets",
+      "dynamicId": null
+    }
+  }
+}
