+++ src/PRN212.AIStudyHub.WebAPI/Properties/serviceDependencies.local.json
@@ -1,0 +1,8 @@
+{
+  "dependencies": {
+    "secrets1": {
+      "type": "secrets.user",
+      "dynamicId": null
+    }
+  }
+}
