+++ src/PRN212.AIStudyHub.WebAPI/appsettings.Development.json
@@ -1,0 +1,23 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=localhost;Database=AIStudyHubDB;User Id=your_user_id;Password=your_password;Trusted_Connection=True;TrustServerCertificate=True;"
+  },
+  "Cloudinary": {
+    "CloudName": "your_cloud_name",
+    "ApiKey": "your_api_key",
+    "ApiSecret": "your_api_secret",
+    "Folder": "ai-study-hub/documents"
+  },
+  "JwtSettings": {
+    "Issuer": "your_issuer",
+    "Audience": "your_audience",
+    "ExpiryInMinutes": 60
+  }
+}
